# docs/swagger.yaml
basePath: /api
definitions:
  models.Attachment:
    properties:
      fileName:
        type: string
      id:
        type: string
      url:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Filed Upload Service that stores files associated by reference id
  title: FileRequest Upload Service
  version: "1.0"
paths:
  /files/:
    post:
      consumes:
      - multipart/form-data
      description: Upload file and attach it to the object of id
      operationId: upload-file
      parameters:
      - description: FileDto
        in: formData
        name: file
        required: true
        type: file
      - description: Reference Object ID
        in: formData
        name: referenceObjectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Upload file
      tags:
      - files
  /files/{id}:
    delete:
      consumes:
      - application/json
      description: Remove file by id
      operationId: remove-file-by-id
      parameters:
      - description: FileID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Remove file
      tags:
      - files
  /files/ping:
    get:
      consumes:
      - application/json
      description: Returns Pong when service is alive
      operationId: files-ping-pong
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Ping-pong healthcheck
      tags:
      - files
  /files/reference/{id}:
    get:
      consumes:
      - application/json
      description: Get file by reference id
      operationId: get-file-by-reference-id
      parameters:
      - description: FileDto ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Attachment'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Show files
      tags:
      - files
swagger: "2.0"